requirements.txt
Flask==2.1.2
flask_cors==3.0.10
mysql-connector-python
pydub==0.25.1
requests==2.25.1
sslkeylog==0.4.0
